Started unit tests. Updated Makefile, set TESTS=1.
Wrote first unit test, and it fails, so need to find out why before I do
anything else. All known bugs are to be fixed before any new code, or
refactoring takes place.

The Makefile is a bit hacky, need to clean that up, Makefiles are... fun.
The unit testing requires CUnit to be installed. As with SDL and ncurses,
pkg-config is used for CFLAGS and LIBS.