Starting refactor of creature functions.
Planning on refacting the game initialisation code. Looks like the player
initialisation code might be easy to pull out. This might not be the case.
However, have found that the program will segfault on OSX using the SDL
renderer. The problem doesn't show up with the GL renderer, but looks to be
related to the "freeGrid" function.

Commiting this code now, so I can work finding the error.